+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+INCOMING
+----
+
+**Inbox**
+- Added `androidCustomLargeIcon` property to `IInboxNotification`.
+- Added `androidBigPicture` property to `IInboxNotification`.
+- Added `deeplink` property to `IInboxNotification`.
+
 12.0.0
 ----